requirements.txt
# Core dependencies
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.66.0

# Evaluation frameworks
evalplus>=0.2.0

# Code similarity and metrics
python-Levenshtein>=0.21.0

# Data processing
jsonlines>=4.0.0

# Optional: for better error handling and retry logic
tenacity>=8.2.0
